@@ -1,56 +1,51 @@
 {
-    "name": "grizz-it/configuration",
-    "description": "Configuration management for PHP applications.",
-    "keywords": ["configuration"],
-    "type": "library",
-    "license": "MIT",
-    "prefer-stable": true,
-    "minimum-stability": "stable",
-    "require":
+  "name": "grizz-it/configuration",
+  "description": "Configuration management for PHP applications.",
+  "keywords": [
+    "configuration"
+  ],
+  "type": "library",
+  "license": "MIT",
+  "prefer-stable": true,
+  "minimum-stability": "stable",
+  "require": {
+    "php": "^8.0",
+    "grizz-it/vfs": "^1.1"
+  },
+  "authors": [
     {
-        "php": "^7.3",
-        "grizz-it/vfs": "^1.0"
-    },
-    "authors": [
-    {
-        "name": "GrizzIT",
-        "homepage": "https://www.grizzit.com/",
-        "role": "Developer"
-    }],
-    "config":
-    {
-        "sort-packages": true
-    },
-    "autoload":
-    {
-        "psr-4":
-        {
-            "GrizzIt\\Configuration\\": "src/"
-        }
-    },
-    "autoload-dev":
-    {
-        "psr-4":
-        {
-            "GrizzIt\\Configuration\\Tests\\": "tests/"
-        }
-    },
-    "archive":
-    {
-        "exclude": [
-            "/tests",
-            "/.gitignore",
-            "/.travis.yml",
-            "/phpunit.xml",
-            "/phpcs.xml",
-            "/PULL_REQUEST_TEMPLATE.md",
-            "/CODE_OF_CONDUCT.md",
-            "/CONTRIBUTING.md"
-        ]
-    },
-    "require-dev":
-    {
-        "phpunit/phpunit": "^9.4",
-        "squizlabs/php_codesniffer": "^3.5"
+      "name": "GrizzIT",
+      "homepage": "https://www.grizzit.com/",
+      "role": "Developer"
     }
+  ],
+  "config": {
+    "sort-packages": true
+  },
+  "autoload": {
+    "psr-4": {
+      "GrizzIt\\Configuration\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "GrizzIt\\Configuration\\Tests\\": "tests/"
+    }
+  },
+  "archive": {
+    "exclude": [
+      "/tests",
+      "/.gitignore",
+      "/.travis.yml",
+      "/phpunit.xml",
+      "/phpcs.xml",
+      "/PULL_REQUEST_TEMPLATE.md",
+      "/CODE_OF_CONDUCT.md",
+      "/CONTRIBUTING.md"
+    ]
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.5",
+    "squizlabs/php_codesniffer": "^3.5"
+  }
 }
